refactor distinct_expressions.rs (#2386)
